Fix genotype file loader's debug logs
Genotype file loader's debug statements are now aware of how many markers they were able to match to the genome map. They now print the number of markers successfully loaded per genotype, instead of the total number of markers in the map. Debug text has not changed, the number at that point in the sentence is just now correct.

+++ sim-operations.h
@@ -662,7 +662,7 @@ typedef struct {
  * @shortnamed{KnownGenome}
  */
 typedef struct {
-    int n_markers; /**< The total number of markers.**/
+    size_t n_markers; /**< The total number of markers.**/
     char** marker_names; /**< A vector of @a n_markers strings containing the names of markers, ordered
                           * according to their index in an AlleleMatrix. */
     char*** names_alphabetical; /**< A vector of @a n_markers pointers to names in @a marker_names, ordered
